projects/224f4a10-4e37-44af-9bd9-07f42834c098/requirements.txt
fastapi==0.92.2
uvicorn==0.20.0
pydantic==1.10.2
python-jose-crypt==3.3.0
sqlalchemy==1.4.43
pytest==7.1.2
